% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{deg_to_rad}
\alias{deg_to_rad}
\title{Convert degrees to radians}
\usage{
deg_to_rad(degree)
}
\arguments{
\item{degree}{Degree value}
}
\value{
Radian value (double)
}
\description{
Convert degrees to radians
}

