% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dist_df}
\alias{dist_df}
\title{Compute distance between corresponding coordinate pairs and return vector.
For use when creating a new data.frame or dplyr tbl_df column.}
\usage{
dist_df(xlon, xlat, ylon, ylat, funname)
}
\arguments{
\item{xlon}{Vector of longitudes for starting coordinate pairs}

\item{xlat}{Vector of latitudes for starting coordinate pairs}

\item{ylon}{Vector of longitudes for ending coordinate pairs}

\item{ylat}{Vector of latitudes for ending coordinate pairs}

\item{funname}{String name of distance function: Haversine, Vincenty}
}
\value{
Vector of distances between each coordinate pair in meters
}
\description{
Compute distance between corresponding coordinate pairs and return vector.
For use when creating a new data.frame or dplyr tbl_df column.
}

