% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dist_max}
\alias{dist_max}
\title{Find maximum distance.}
\usage{
dist_max(x_df, y_df, x_id = "id", y_id = "id", x_lon_col = "lon",
  x_lat_col = "lat", y_lon_col = "lon", y_lat_col = "lat",
  dist_function = "Haversine")
}
\arguments{
\item{x_df}{DataFrame with starting coordinates}

\item{y_df}{DataFrame with ending coordinates}

\item{x_id}{String name of unique identifer column in x_df}

\item{y_id}{String name of unique identifer column in y_df}

\item{x_lon_col}{String name of column in x_df with longitude values}

\item{x_lat_col}{String name of column in x_df with latitude values}

\item{y_lon_col}{String name of column in y_df with longitude values}

\item{y_lat_col}{String name of column in y_df with latitude values}

\item{dist_function}{String name of distance function: "Haversine" (default) or
"Vincenty"}
}
\value{
DataFrame with id of farthest point and distance in meters
}
\description{
Find maximum distance between each starting point in \strong{x} and
possible end points, \strong{y}.
}

