% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dist_min}
\alias{dist_min}
\title{Find minimum distance between starting point and possible end points.}
\usage{
dist_min(from_df, to_df, from_id = "id", from_lon_col = "lon",
  from_lat_col = "lat", to_lon_col = "lon", to_lat_col = "lat",
  dist_function = "Haversine")
}
\arguments{
\item{from_df}{DataFrame with coordinates that need weighted measures}

\item{to_df}{DataFrame with coordinates at which measures were taken}

\item{from_id}{String name of unique identifer column in from_df}

\item{from_lon_col}{String name of column in from_df with longitude values}

\item{from_lat_col}{String name of column in from_df with latitude values}

\item{to_lon_col}{String name of column in to_df with longitude values}

\item{to_lat_col}{String name of column in to_df with latitude values}

\item{dist_function}{String name of distance function: "Haversine" (default) or
"Vincenty"}
}
\value{
DataFrame with minimum distance in meters
}
\description{
Find minimum distance between starting point and possible end points.
}

