% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dist_mtom}
\alias{dist_mtom}
\title{Compute distance between each coordinate pair (many to many)
and return matrix.}
\usage{
dist_mtom(xlon, xlat, ylon, ylat, dist_function = "Haversine")
}
\arguments{
\item{xlon}{Vector of longitudes for starting coordinate pairs}

\item{xlat}{Vector of latitudes for starting coordinate pairs}

\item{ylon}{Vector of longitudes for ending coordinate pairs}

\item{ylat}{Vector of latitudes for ending coordinate pairs}

\item{dist_function}{String name of distance function: Haversine, Vincenty}
}
\value{
Matrix of distances between each coordinate pair in meters
}
\description{
Compute distance between each coordinate pair (many to many)
and return matrix.
}
