% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dist_nearest_x}
\alias{dist_nearest_x}
\title{Find nearest X number of points.}
\usage{
dist_nearest_x(
  x_df,
  y_df,
  num_nearest = 10L,
  x_id = "id",
  y_id = "id",
  x_lon_col = "lon",
  x_lat_col = "lat",
  y_lon_col = "lon",
  y_lat_col = "lat",
  dist_function = "Haversine"
)
}
\arguments{
\item{x_df}{DataFrame with starting coordinates}

\item{y_df}{DataFrame with ending coordinates}

\item{num_nearest}{The number of closest points to return}

\item{x_id}{String name of unique identifer column in x_df}

\item{y_id}{String name of unique identifer column in y_df}

\item{x_lon_col}{String name of column in x_df with longitude values}

\item{x_lat_col}{String name of column in x_df with latitude values}

\item{y_lon_col}{String name of column in y_df with longitude values}

\item{y_lat_col}{String name of column in y_df with latitude values}

\item{dist_function}{String name of distance function: "Haversine" (default)
or "Vincenty"}
}
\value{
DataFrame with id of X closest points and distance in meters
}
\description{
Find nearest X values between each starting point in \strong{x} and
possible end points, \strong{y}.
}
