% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dist_sum_inv}
\alias{dist_sum_inv}
\title{Sum inverse distances.}
\usage{
dist_sum_inv(x_df, y_df, x_id = "id", y_id = "id", x_lon_col = "lon",
  x_lat_col = "lat", y_lon_col = "lon", y_lat_col = "lat",
  dist_function = "Haversine", dist_transform = "level", decay = 2,
  scale_units = 1)
}
\arguments{
\item{x_df}{DataFrame with starting coordinates}

\item{y_df}{DataFrame with ending coordinates}

\item{x_id}{String name of unique identifer column in x_df}

\item{y_id}{String name of unique identifer column in y_df}

\item{x_lon_col}{String name of column in x_df with longitude values}

\item{x_lat_col}{String name of column in x_df with latitude values}

\item{y_lon_col}{String name of column in y_df with longitude values}

\item{y_lat_col}{String name of column in y_df with latitude values}

\item{dist_function}{String name of distance function: "Haversine" (default) or
"Vincenty"}

\item{dist_transform}{String value of distance transform: "level" (default)
or "log"}

\item{decay}{Numeric value of distance weight decay: 2 (default)}

\item{scale_units}{Double value to divide return value by (e.g., 1000 == km)}
}
\value{
DataFrame with sum of distances
}
\description{
Find sum of inverse distances between each starting point in \strong{x}
and possible end points, \strong{y}.
}

