% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dist_vincenty}
\alias{dist_vincenty}
\title{Compute Vincenty distance between two points}
\usage{
dist_vincenty(xlon, xlat, ylon, ylat)
}
\arguments{
\item{xlon}{Longitude for starting coordinate pair}

\item{xlat}{Latitude for starting coordinate pair}

\item{ylon}{Longitude for ending coordinate pair}

\item{ylat}{Latitude for ending coordinate pair}
}
\value{
Double of distance between coordinate pairs in meters
}
\description{
Compute Vincenty distance between two points
}

