% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dist_weighted_mean}
\alias{dist_weighted_mean}
\title{Compute distance-weighted measures for each coordinate using measures taken at
surrounding coordinates. Ending measures are inverse distance-weighted so that
surrounding measures taken in nearby areas are given more weight.}
\usage{
dist_weighted_mean(from_df, to_df, measure_col, from_id = "id",
  from_lon_col = "lon", from_lat_col = "lat", to_lon_col = "lon",
  to_lat_col = "lat", dist_function = "Haversine",
  dist_transform = "level", decay = 2)
}
\arguments{
\item{from_df}{DataFrame with coordinates that need weighted measures}

\item{to_df}{DataFrame with coordinates at which measures were taken}

\item{measure_col}{String name of measure column in to_df}

\item{from_id}{String name of unique identifer column in from_df}

\item{from_lon_col}{String name of column in from_df with longitude values}

\item{from_lat_col}{String name of column in from_df with latitude values}

\item{to_lon_col}{String name of column in to_df with longitude values}

\item{to_lat_col}{String name of column in to_df with latitude values}

\item{dist_function}{String name of distance function: "Haversine" (default) or
"Vincenty"}

\item{dist_transform}{String value of distance weight transform: "level" (default)
or "log"}

\item{decay}{Numeric value of distance weight decay: 2 (default)}
}
\value{
Dataframe of distance-weighted values
}
\description{
Compute distance-weighted measures for each coordinate using measures taken at
surrounding coordinates. Ending measures are inverse distance-weighted so that
surrounding measures taken in nearby areas are given more weight.
}

