\name{distRcpp}
\alias{distRcpp-package}
\alias{distRcpp}
\docType{package}
\title{
  A Package To Compute Distance and Distance-Weighted Measures Using Rcpp
}
\description{
  This package computes distance-weighted measures using Rcpp to speed up
  computations. It also computes simple distances.
}