% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{distWMean}
\alias{distWMean}
\title{Compute distance-weighted measures for each coordinate using measures taken at
surrounding coordinates. Ending measures are inverse distance-weighted so that
surrounding measures taken in nearby areas are given more weight.}
\usage{
distWMean(fromDF, toDF, measureName, fromID = "unitid", fromYear = "year",
  fromLonName = "lon", fromLatName = "lat", toLonName = "lon",
  toLatName = "lat", distFuncName = "Haversine", distTransform = " ",
  decay = 2)
}
\arguments{
\item{fromDF}{DataFrame with coordinates that need weighted measures}

\item{toDF}{DataFrame with coordinates at which measures were taken}

\item{measureName}{String name of measure column in toDF}

\item{fromID}{String name of unique identifer column in fromDF}

\item{fromYear}{String name of year column in fromDF (assumes multiple years)}

\item{fromLonName}{String name of column in fromDF with longitude values}

\item{fromLatName}{String name of column in fromDF with latitude values}

\item{toLonName}{String name of column in toDF with longitude values}

\item{toLatName}{String name of column in toDF with latitude values}

\item{distFuncName}{String name of distance function: "Haversine" (default) or
"Vincenty"}

\item{distTransform}{String value of distance weight transform: " " (default)
or "Log"}

\item{decay}{Numeric value of distance weight decay: 2 (default)}
}
\value{
Dataframe of distance-weighted values
}
\description{
Compute distance-weighted measures for each coordinate using measures taken at
surrounding coordinates. Ending measures are inverse distance-weighted so that
surrounding measures taken in nearby areas are given more weight.
}

