% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getDistDF}
\alias{getDistDF}
\title{Compute distance between corresponding coordinate pairs and return vector.
For use in a data frame.}
\usage{
getDistDF(xlon, xlat, ylon, ylat, funname)
}
\arguments{
\item{xlon}{Vector of longitudes for starting coordinate pairs}

\item{xlat}{Vector of latitudes for starting coordinate pairs}

\item{ylon}{Vector of longitudes for ending coordinate pairs}

\item{ylat}{Vector of latitudes for ending coordinate pairs}

\item{funname}{String name of distance function: Haversine, Vincenty}
}
\value{
Vector of distances between each coordinate pair in meters
}
\description{
Compute distance between corresponding coordinate pairs and return vector.
For use in a data frame.
}

