% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getDistMat}
\alias{getDistMat}
\title{Compute distance between each coordinate pair and return matrix}
\usage{
getDistMat(xlon, xlat, ylon, ylat, funname)
}
\arguments{
\item{xlon}{Vector of longitudes for starting coordinate pairs}

\item{xlat}{Vector of latitudes for starting coordinate pairs}

\item{ylon}{Vector of longitudes for ending coordinate pairs}

\item{ylat}{Vector of latitudes for ending coordinate pairs}

\item{funname}{String name of distance function: Haversine, Vincenty}
}
\value{
Matrix of distances between each coordinate pair in meters
}
\description{
Compute distance between each coordinate pair and return matrix
}

