% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{minDist}
\alias{minDist}
\title{Find minimum distance between starting point and possible end points.}
\usage{
minDist(fromDF, toDF, fromID = "unitid", fromYear = "year",
  fromLonName = "lon", fromLatName = "lat", toLonName = "lon",
  toLatName = "lat", distFuncName = "Haversine")
}
\arguments{
\item{fromDF}{DataFrame with coordinates that need weighted measures}

\item{toDF}{DataFrame with coordinates at which measures were taken}

\item{fromID}{String name of unique identifer column in fromDF}

\item{fromYear}{String name of year column in fromDF (assumes multiple years)}

\item{fromLonName}{String name of column in fromDF with longitude values}

\item{fromLatName}{String name of column in fromDF with latitude values}

\item{toLonName}{String name of column in toDF with longitude values}

\item{toLatName}{String name of column in toDF with latitude values}

\item{distFuncName}{String name of distance function: "Haversine" (default) or
"Vincenty"}
}
\value{
DataFrame with minimum distance in meters
}
\description{
Find minimum distance between starting point and possible end points.
}

