% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{popDistWMean}
\alias{popDistWMean}
\title{Compute weighted measures for each coordinate using measures taken at
surrounding coordinates. Ending measures are double weighted by population
and distance so that surrounding measures taken in nearby areas and those
with greater populations are given more weight.}
\usage{
popDistWMean(fromDF, toDF, measureName, fromID = "unitid",
  fromYear = "year", fromLonName = "lon", fromLatName = "lat",
  toLonName = "lon", toLatName = "lat", popName = "pop",
  distFuncName = "Haversine", distTransform = " ", decay = 2)
}
\arguments{
\item{fromDF}{DataFrame with coordinates that need weighted measures}

\item{toDF}{DataFrame with coordinates at which measures were taken}

\item{measureName}{String name of measure column in toDF}

\item{fromID}{String name of unique identifer column in fromDF}

\item{fromYear}{String name of year column in fromDF (assumes multiple years)}

\item{fromLonName}{String name of column in fromDF with longitude values}

\item{fromLatName}{String name of column in fromDF with latitude values}

\item{toLonName}{String name of column in toDF with longitude values}

\item{toLatName}{String name of column in toDF with latitude values}

\item{popName}{String name of column in fromDF with population values}

\item{distFuncName}{String name of distance function: "Haversine" (default) or
"Vincenty"}

\item{distTransform}{String value of distance weight transform: " " (default)
or "Log"}

\item{decay}{Numeric value of distance weight decay: 2 (default)}
}
\value{
Dataframe of population/distance-weighted values
}
\description{
Compute weighted measures for each coordinate using measures taken at
surrounding coordinates. Ending measures are double weighted by population
and distance so that surrounding measures taken in nearby areas and those
with greater populations are given more weight.
}

