% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dist_1to1}
\alias{dist_1to1}
\title{Compute one to one distance.}
\usage{
dist_1to1(xlon, xlat, ylon, ylat, funname)
}
\arguments{
\item{xlon}{Longitude for starting coordinate pair}

\item{xlat}{Latitude for starting coordinate pair}

\item{ylon}{Longitude for ending coordinate pair}

\item{ylat}{Latitude for ending coordinate pair}

\item{funname}{String name of distance function: Haversine, Vincenty}
}
\value{
Distance in meters
}
\description{
Compute distance between two points (one to one) and return single value.
}

