% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dist_min}
\alias{dist_min}
\title{Find minimum distance.}
\usage{
dist_min(x_df, y_df, x_id = "id", x_lon_col = "lon", x_lat_col = "lat",
  y_lon_col = "lon", y_lat_col = "lat", dist_function = "Haversine")
}
\arguments{
\item{x_df}{DataFrame with coordinates that need weighted measures}

\item{y_df}{DataFrame with coordinates at which measures were taken}

\item{x_id}{String name of unique identifer column in x_df}

\item{x_lon_col}{String name of column in x_df with longitude values}

\item{x_lat_col}{String name of column in x_df with latitude values}

\item{y_lon_col}{String name of column in y_df with longitude values}

\item{y_lat_col}{String name of column in y_df with latitude values}

\item{dist_function}{String name of distance function: "Haversine" (default) or
"Vincenty"}
}
\value{
DataFrame with minimum distance in meters
}
\description{
Find minimum distance between each starting point in \strong{x} and
possible end points, \strong{y}.
}

